backend/requirements.txt
# SalonSync Backend Dependencies

# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
gunicorn>=21.0.0
python-multipart>=0.0.6

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
asyncpg>=0.29.0

# Authentication
python-jose[cryptography]>=3.3.0
argon2-cffi>=23.1.0
bcrypt>=4.0.0

# Data Validation
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Background Tasks
celery>=5.3.0
redis>=4.2.0,<5.0.0
apscheduler>=3.10.0

# Email
aiosmtplib>=3.0.0
jinja2>=3.1.0

# HTTP Client
httpx>=0.26.0
requests>=2.31.0

# Date/Time
python-dateutil>=2.8.2
pytz>=2024.1

# Utilities
python-dotenv>=1.0.0
Pillow>=10.2.0

# ═══════════════════════════════════════════════════════════════
# SalonSync-Specific Dependencies
# ═══════════════════════════════════════════════════════════════

# Media Storage (Cloudinary for photos)
cloudinary>=1.38.0

# AI Services
anthropic>=0.18.0

# Payment Processing
stripe>=7.0.0

# SMS Notifications
twilio>=8.0.0

# Social Media APIs
# Instagram Graph API uses httpx (already included)
# TikTok API uses httpx (already included)

# Image Processing (for comparison photos)
opencv-python-headless>=4.9.0

# Slugify for URL-friendly salon names
python-slugify>=8.0.0

# Rate limiting
slowapi>=0.1.9

# CORS
# (included in fastapi)

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0  # for TestClient

# Code Quality (dev)
black>=24.1.0
isort>=5.13.0
flake8>=7.0.0
mypy>=1.8.0

# Type stubs
types-python-dateutil>=2.8.0
types-requests>=2.31.0
types-pytz>=2024.1.0
